Fix an issue where completion buffers where not being cleaned up
When ever the process was canceled it would leave the buffer it was
using lying around. So instead we switch to using get-buffer-create
which will reuse the same buffer everytime. However This also means
that we need to make sure that the buffer-process is dead and the
buffer is empty before we start a new one.