Allow completion in buffers that are not linked to valid files
sometimes when a directory is deleted any buffer that was using that
directory as its default-directory will no longer be able to run
sub-processes because the directory the shell would use is invalid. So
if default-directory is invalid we use user-emacs-directory instead.
Relative paths will be wrong, but at least we have some completion.